+++ sdks/python/tests/dspy_tracking/test_gepa_tracking.py
@@ -7,7 +7,6 @@ import json
 from typing import Any
 
 import dspy
-import httpx
 import pytest
 from dspy.teleprompt import GEPA
 
@@ -23,66 +22,13 @@ from langwatch.dspy import (
     langwatch_dspy,
 )
 
-
-class Program(dspy.Module):
-    def __init__(self):
-        super().__init__()
-        self.answer = dspy.Predict("question -> answer")
-
-    def forward(self, question: str):
-        return self.answer(question=question)
-
-
-def feedback_metric(gold, pred, trace=None, pred_name=None, pred_trace=None):
-    return dspy.Prediction(score=0.5, feedback="too many steps")
-
-
-def build_optimizer(**gepa_kwargs) -> GEPA:
-    return dspy.GEPA(
-        metric=feedback_metric,
-        max_metric_calls=4,
-        reflection_minibatch_size=2,
-        reflection_lm=dspy.LM("openai/gpt-5-mini"),
-        gepa_kwargs=gepa_kwargs or None,
-    )
-
-
-def tracked_optimizer(**gepa_kwargs) -> LangWatchTrackedGEPA:
-    optimizer = build_optimizer(**gepa_kwargs)
-    optimizer.__class__ = LangWatchTrackedGEPA
-    optimizer.patch()  # type: ignore[attr-defined]
-    return optimizer  # type: ignore[return-value]
-
-
-def valset_of(size: int) -> list[dspy.Example]:
-    return [
-        dspy.Example(question=f"question {i}", answer=f"answer {i}").with_inputs(
-            "question"
-        )
-        for i in range(size)
-    ]
-
-
-def valset_event(
-    *,
-    candidate_idx: int,
-    candidate: dict[str, str],
-    scores_by_val_id: dict[int, float],
-    outputs_by_val_id: Any,
-    average_score: float,
-) -> dict[str, Any]:
-    return {
-        "iteration": candidate_idx,
-        "candidate_idx": candidate_idx,
-        "candidate": candidate,
-        "scores_by_val_id": scores_by_val_id,
-        "average_score": average_score,
-        "num_examples_evaluated": len(scores_by_val_id),
-        "total_valset_size": len(scores_by_val_id),
-        "parent_ids": [],
-        "is_best_program": True,
-        "outputs_by_val_id": outputs_by_val_id,
-    }
+from .helpers import (
+    Program,
+    build_optimizer,
+    tracked_optimizer,
+    valset_event,
+    valset_of,
+)
 
 
 @pytest.fixture
@@ -325,78 +271,3 @@ class TestWhenACustomOptimizerLogsAStep:
         assert len(sent_steps[0].examples) == 1
         assert sent_steps[0].examples[0].score == 1.0
         assert langwatch_dspy.examples_buffer == []
-
-
-class Posts:
-    """The step posts made so far, answered with `statuses` in order, then 200."""
-
-    def __init__(self):
-        self.statuses: list[int] = []
-        self.bodies: list[list[dict]] = []
-
-    def post(self, url, **kwargs):
-        self.bodies.append(json.loads(kwargs["data"]))
-        return httpx.Response(
-            self.statuses.pop(0) if self.statuses else 200,
-            request=httpx.Request("POST", url),
-        )
-
-
-@pytest.fixture
-def posts(monkeypatch) -> Posts:
-    """A tracker with a run and no waits between step post retries."""
-    posts = Posts()
-
-    langwatch_dspy.run_id = "run"
-    langwatch_dspy.experiment_slug = "experiment"
-    langwatch_dspy.workflow_version_id = None
-    langwatch_dspy.reset()
-    monkeypatch.setattr(langwatch, "get_api_key", lambda: "key")
-    monkeypatch.setattr(langwatch.dspy.httpx, "post", posts.post)
-    monkeypatch.setattr(langwatch.dspy.time, "sleep", lambda seconds: None)
-    return posts
-
-
-def log_a_step(index: str) -> None:
-    langwatch_dspy.log_step(
-        optimizer=DSPyOptimizer(name="GEPA", parameters={}),
-        index=index,
-        score=1.0,
-        label="score",
-        predictors=[],
-    )
-
-
-class TestWhenTheStepPostFails:
-    # @scenario "A step that could not be sent goes out with the next one"
-    def test_resends_the_step_with_the_next_one(self, posts):
-        posts.statuses.extend([502, 502, 502, 502, 502])
-
-        with pytest.raises(Exception):
-            log_a_step("0")
-
-        assert len(posts.bodies) == 5
-        assert [step.index for step in langwatch_dspy.steps_buffer] == ["0"]
-
-        log_a_step("1")
-
-        assert [step["index"] for step in posts.bodies[-1]] == ["0", "1"]
-        assert langwatch_dspy.steps_buffer == []
-
-    # @scenario "The steps a failed post left behind are sent when the run ends"
-    def test_flushes_the_buffer_when_the_run_ends(self, posts):
-        posts.statuses.extend([502, 502, 502, 502, 502])
-        with pytest.raises(Exception):
-            log_a_step("0")
-        callback = LangWatchGEPACallback(
-            optimizer=tracked_optimizer(), student=Program(), valset=valset_of(1)
-        )
-
-        callback.on_optimization_end({})  # type: ignore[arg-type]
-
-        assert [step["index"] for step in posts.bodies[-1]] == ["0"]
-        assert langwatch_dspy.steps_buffer == []
-
-        callback.on_optimization_end({})  # type: ignore[arg-type]
-
-        assert len(posts.bodies) == 6
